Add posix platform
